+++ lib/oxidized/model/comware.rb
@@ -56,8 +56,9 @@ class Comware < Oxidized::Model
         cmd vars(:comware_cmdline)
 
         # HP V1950 OS r3208 (v7.1)
-        cmd 'xtd-cli-mode', /(#{@node.prompt}|Continue)/
-        cmd 'y', /(#{@node.prompt}|input password)/
+        # HPE Office Connect 1950
+        cmd 'xtd-cli-mode', /(#{@node.prompt}|Continue|Switch)/
+        cmd 'y', /(#{@node.prompt}|input password|Password:)/
         cmd vars(:comware_cmdline)
       end
     end