+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Github Action: Support creating custom version tag in a release
 - Add support for no enable password set on ironware
 - change pfSense secret scrubbing to keep config as well-formed XML
+- Add support for comware HPE Office Connect 1950
 
 ### Added
 
@@ -51,6 +52,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - model for YAMAHA NVR/RTX Series (@bluekirin55)
 - model for ZPE Nodegrid OS (@euph333)
 - model for H3C switches
+- model for Cisco Catalyst Express switches (@unemongod)
 - extended mysql source configuration to include tls options (@glaubway)
 - updated rugged in gemspec for ruby 3.0 support (@firefishy)